requirements.txt
PyMuPDF 
Pillow 
numpy 
scikit-image